move HandleBox to finish_handles.hpp
