+++ include/condy/awaiters.hpp
@@ -18,39 +18,10 @@
 #include "condy/work_type.hpp"
 #include <coroutine>
 #include <cstddef>
-#include <memory>
 #include <stdexcept>
 #include <tuple>
 
 namespace condy {
-
-template <OpFinishHandleLike Handle> class HandleBox {
-public:
-    HandleBox(Handle h) : handle_(std::move(h)) {}
-
-    Handle &get() noexcept { return handle_; }
-
-    void maybe_release() noexcept { /* No-op */ }
-
-private:
-    Handle handle_;
-};
-
-template <CQEHandlerLike CQEHandler, typename Func>
-class HandleBox<ZeroCopyOpFinishHandle<CQEHandler, Func>> {
-public:
-    using Handle = ZeroCopyOpFinishHandle<CQEHandler, Func>;
-    HandleBox(Handle h) : handle_ptr_(std::make_unique<Handle>(std::move(h))) {}
-    HandleBox(const HandleBox &other) // Deep copy
-        : handle_ptr_(std::make_unique<Handle>(*other.handle_ptr_)) {}
-
-    Handle &get() noexcept { return *handle_ptr_; }
-
-    void maybe_release() noexcept { handle_ptr_.release(); }
-
-private:
-    std::unique_ptr<Handle> handle_ptr_;
-};
 
 template <OpFinishHandleLike Handle, PrepFuncLike Func> class OpAwaiterBase {
 public:
